+++ src/main/java/org/example/DictionaryRepository/ConsoleDictionaryRepository.java
@@ -6,15 +6,13 @@ public class ConsoleDictionaryRepository implements IDictionaryRepository {
 
     /**
      * Вывродит сожержимое словаря в консорль
-     *
-     * @param source     адрес сохранёной страницы
+     *  @param source     адрес сохранёной страницы
      * @param dictionary словарь, содержащий слова и частоту их
-     *                  нахождения на странице
      */
     @Override
-    public void store(String source, Map<String, Integer> dictionary) {
+    public void store(String source, Map<String, Long> dictionary) {
         System.out.println("Data from " + source + ":");
-        for (Map.Entry<String, Integer> item : dictionary.entrySet()) {
+        for (Map.Entry<String, Long> item : dictionary.entrySet()) {
             System.out.println(item.getKey() + "\t" + item.getValue());
         }
     }
